fix tests

+++ tests/unit/butterfree/core/load/conftest.py
@@ -4,6 +4,7 @@ from pytest import fixture
 from butterfree.core.configs.db import CassandraConfig
 from butterfree.core.constants import columns
 from butterfree.core.constants.columns import TIMESTAMP_COLUMN
+from butterfree.core.constants.data_type import DataType
 from butterfree.core.transform import FeatureSet
 from butterfree.core.transform.features import Feature, KeyFeature, TimestampFeature
 
@@ -13,7 +14,7 @@ def feature_set():
     key_features = [KeyFeature(name="id", description="Description")]
     ts_feature = TimestampFeature(from_column=TIMESTAMP_COLUMN)
     features = [
-        Feature(name="feature", description="Description"),
+        Feature(name="feature", description="Description", dtype=DataType.BIGINT,)
     ]
     return FeatureSet(
         "feature_set",
